pyproject.toml
[tool.poetry]
name = "loyalt"
version = "1.0.0"
description = "LoaylT"
authors = ["AKRBOUT"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10.5"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
pillow = "^10.2.0"
sqlalchemy="2.0.38"
validators="0.34.0"
email-validator="2.2.0"
qrcode="8.0"
pydantic="2.10.6"
pyjwt="2.10.1"
passlib="1.7.4"
psycopg="3.2.5"
psycopg-binary="3.2.5"
aiobotocore="2.19.0"
python-multipart="0.0.20"
argon2-cffi="23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"